Fixed Issue of Displaying of Incorrect Data in Editing Modal Windows
Fixed issue of incorrect data being shown in EditForms of modal windows for editing of addresses and email addresses when there are several records in a database in the appropriate tables (for instance, if you have two records describing email addresses of some leaseholder and pressed button to invoke modal window allowing to edit second email address then did some editing and cancelled it by pressing of X symbol or by pressing on a background then next time you pressed Edit button you saw value stored in a database of a first email address stored in a database in a field for editing). Now displaying of information works correctly and shows correct data of appropriate address or email address in editing field of appropriate modal window.